chore: rename vite base config helper
