+++ vite.base.ts
