+++ vite.config.ts
@@ -5,7 +5,7 @@ import { defineConfig, mergeConfig } from 'vite';
 
 import { viteElectronAppPlugin } from './infrastructure/viteElectronAppPlugin';
 import { version } from './package.json';
-import { external, getBuildConfig } from './vite.base.config';
+import { external, getBuildConfig } from './vite.base';
 
 // https://vitejs.dev/config
 export default defineConfig((env) => {
